Fix and optimize the squared absolute value
The function for calculating the square absolute value was relatively
slow, as it used the slower np.square and np.abs procedures. Additionally,
the docstrings for cabs2 and abs2 functions were misleading: one could
think that the former operates on complex arrays only, and the latter -
on real arrays only, which is not the case: they both operate on both
types of arrays, but cabs2 just returns a complex result if the argument
is complex.

Also, a case of unnecessary calling of cabs2 instead of abs2 is fixed.

This patch increases the overall performance speed slightly, because of
the faster performance of abs2.

Signed-off-by: Leonid Bloch <leonid.bloch@esrf.fr>