Implement shift in diversify function
This allows probes/objects with more than one mode to receive
2d shifts applied to each mode which can be defined relatively
with a single tuple, such that modes are placed equally at
positive and negative multiple of this tuple, or by specifying
an individual tuple for each mode.

The shift is implemented with the shift_zoom function from
array_utils.

To match behaviour of other diversity parameters the first layer
always receives no shift.